+++ client/src/client.rs
@@ -203,7 +203,7 @@ impl Client {
     }
 
     /// Validates if a bech32 HRP matches the one from the connected network.
-    pub async fn validate_bech32_hrp(&self, bech32_hrp: &str) -> Result<()> {
+    pub async fn bech32_hrp_matches(&self, bech32_hrp: &str) -> Result<()> {
         let expected_bech32_hrp = self.get_bech32_hrp().await?;
         if bech32_hrp != expected_bech32_hrp {
             return Err(Error::InvalidBech32Hrp {