+++ wallet/src/account/operations/transaction/high_level/create_alias.rs
@@ -118,7 +118,7 @@ impl AccountHandle {
         {
             Some(bech32_address) => {
                 let (address, bech32_hrp) = Address::try_from_bech32_with_hrp(bech32_address)?;
-                self.client.validate_bech32_hrp(&bech32_hrp).await?;
+                self.client.bech32_hrp_matches(&bech32_hrp).await?;
                 address
             }
             None => {