+++ wallet/src/account/operations/transaction/high_level/send_nft.rs
@@ -73,7 +73,7 @@ impl AccountHandle {
 
         for address_and_nft_id in addresses_nft_ids {
             let (address, bech32_hrp) = Address::try_from_bech32_with_hrp(address_and_nft_id.address)?;
-            self.client.validate_bech32_hrp(&bech32_hrp).await?;
+            self.client.bech32_hrp_matches(&bech32_hrp).await?;
 
             // Find nft output from the inputs
             if let Some(nft_output_data) = unspent_outputs.iter().find(|o| {